requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
python-jose
passlib[bcrypt]
pydantic
pydantic-settings
redis
apscheduler
oss2
